Language support tables + base functionality
Creation of new file for base code for language changing functionality. Will be changed based on other aspects of interface which are yet to be implemented. Language translation tables (mostly) complete, will complete these over the coming weekend. Added js file should provide error if a translation for a key cannot be found